Star Shooter
A space shooter game in which the player shoots as many asteroids that project outwards from the screen as they game. Within a given timer, the player is scored based on their performance. This game was created on the processing platform with java using objected-oriented programming and processing Java libraries.